requirements.txt
chess
pillow
cairosvg
pandas
reportlab
